Improved management of the GUI -> Hub websocket
The GUI now passes a callback id to the server when issuing
a request.
The server sends back the id with the result of the request.
This allows to implement a simple callback mechanism on the
client side (in basic_webapp/websocket.js, and sample use
in basic_webapp/main.js).