Use plain socket communication for hub <-> daemons
It was previously based on json over a websocket
(same as hub <-> web-gui).
We now use pickle over plain sockets for better efficiency.

Note: hub/gui subdirectory was renamed to hub/web in order
to reflect the fact that the whole web management of the hub
is implemented in this directory (not only code strongly related
to the client-side).