GUI -> Hub: add concurrency through parallel websockets
In workflow/js/websocket.js, if a 2nd request is sent on
the websocket using ws_request() while the previous one
did not get back its result yet, we open another websocket
connection to the hub and send the request there.

In order to be even more generic, we maintain a table of
'free' websockets (those that already got the last request
result back), and automatically open a new one when a new
request is issued and this table is empty.

On the hub side, this allows to handle concurrent requests
with no more changes in the code.

In this commit however I also renamed the 'gui_' prefix
to 'rpc_', in order to make it more explicit (especially
if we need other types of connections gui <-> hub in the
future).