CodeOperator: add a tab to demonstrate API usage
This tab "API-test" shows a tree of the operator files
and the content of its icon.svg file.

The code of this tab is made of files testtab.html and
testtab.js.

testtab.js uses API calls to the hub in order to
get appropriate data. 2 calls are made in this code:
* sakura.operator.get_file_tree(<callback>)
* sakura.operator.get_file_content(<file_path>, <callback>)

<callback> must be a function that will be called when
the hub returns the result. The result will be passed
as its first parameter.

A few more API calls related to operator files are provided.
See workflow/js/sakura-operator.js.