Map operator: compute & display a heatmap.
The daemon (map/operator.py) computes the number of points
in each 5x5px cell of the map.
This data (non-zero cells with their coordinates and number
of points) is sent to the UI in a compressed way (we send
integers with appropriate scaling factors and offsets).
On the UI-side, leaflet.heat is used to display the heatmap
given these pre-computed data.