DB objects: Use same names in datastore and as metadata
In a datastore (e.g. a postgresql service),
we can use arbitrary characters in names of
databases, tables and columns as long as we
escape them using double-quotes in queries
(i.e. using "<name>").
Before, we were managing 2 different names,
one was the user-provided name and the other one
was the real object name in the datastore.
By escaping queries appropriately, this commit
allows to use only one name and thus the code
is much simplified.