Hub can now return error messages to GUI.
Messages returned to the GUI are now with the form:
(<req_id>, (<bool>, <value>))
<req_id> allows the GUI to know which request this
response is about.
<bool> is True if everything went well, and <value>
gives the result of the request.
if the hub detected a problem in the request (e.g.
the request was about registering a new user but the
given login already exists), it may return
<bool> = false and <value> = <error_message> instead.
All this is handled in sakura.common.ws_request(),
this the existing calling code is still compatible.
However, the calling code may now add a parameter
to specify the callback that will be called in case
of error.
If this callback is not specified, the error message
is displayed using alert(<msg>).