Session management.
Hub side: websockets are now associated with a given session.
There are 2 entrypoint URLs, one which will create a websocket
associated to a new session, another one which will connect
the websocket to an existing session.
In order to connect to an existing session, one must call
the API function generate_session_secret() on an already
existing websocket connected to the session, and use this
secret in the URL.
The hub now has one more greenlet for maintenance operations.
For now, this greenlet deletes sessions which are not used
anymore.

On the javascript side, ws_request() was updated to connect
all websockets to the same session. However, code in an iframe
still opens a new session. This will have to be fixed in a
later commit.