Web-Cache mechanism
We can now use a web cache mechanism to access javascript
libraries or css files through the hub.

Given the fact that our working directory is set at
$HOME/.sakura when using run-<something>.sh dev scripts,
the web cache directory is at: "$HOME/.sakura/webcache/".

If we target the following URL
http://localhost:8081/webcache/cdnjs/twitter-bootstrap/3.3.7/css/bootstrap-theme.css
the hub will check if the file exists in the cache at
$HOME/.sakura/webcache/cdnjs/twitter-bootstrap/3.3.7/css/bootstrap-theme.css
If not, it will download it from
https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/3.3.7/css/bootstrap-theme.css
and store it there.
Then, the hub will respond to the HTTP request by serving this
file in the cache.

This mechanism allows:
* To download those files only once, thus the system will continue
  to work if an internet connection issue happens
* To circumvent CORS issues (from the browser point of view, all files
  are served from the hub)
* To avoid redistributing code of libraries within our own source code
  (which could possibly cause licensing issues)
* Maybe to increase performance a bit

For now html files were not modified to use web-cache URLs.
This will be done in a later commit.
And only one CDN is allowed: cdnjs. Others could be handled in the
future if needed.