Fix thread identification issue with mac os x
We were using threading.get_ident() but on mac os x
it seams that processes return the same value (may be
a startup timestamp, it seems...).
We now use os.getpid(). We do not use threads at this
time anyway, so that is enough. If we ever use threads
in the future, that would be updated again.