Unload operator tab iframes on modal close.
Src attribute of iframe is set to the empty string
when operator modal window is closed.
This allows to stop any javascript processing occuring
in these iframes and probably not needed anymore since
the operator is not displayed.
An example in the basic-ogl operator: when closing the
modal, the iframe is released, thus the motion-jpeg
stream is closed, thus the OpenGL code running on the
backend is stopped.