w3IncludeHtml: fix sync-mode when loading divs
Previously, all divs were loaded in "sync mode", one after the other.
This hack had been introduced in w3data.js in order to support permalinks.
Because objects referenced by the permalink are not known yet when the
application is starting, this hack was ensuring w3IncludeHtml() is
running in sync-mode, and once done, we are sure the object referenced
by the permalink can be accessed.

This commit restores async-mode, but allows to give w3IncludeHtml() a
callback which will be called when the whole application has been
loaded. In the case of a permalink, this callback is responsible for
loading the div referenced.