New command sakura-sandbox
Command sakura-sandbox allows to ease sakura operator development. There
is no more a need for installing sakura platform components (hub &
daemons). Instead, one just uses the tool sakura-sandbox, which connects
to an existing sakura platform, already in place. Then the tool allows
to expose a directory of one's the development machine. It will be
explored in search for one's operator being developped, and then this
operator appears on the web interface of the platform. If one makes
code changes and hits the 'reload' button of the operator, those changes
will be immediately taken into account.

This commit provides basic functionality for this feature.