Improve standard streams redirection mechanism.
A module attribute '__streams__'  is now added by the import system
in order to link source code to a given redirection.

When interacting with standard streams, we now look for this optional
attribute in the backtrace.

If an operator is instanciated from sakura-shell, standard streams are
redirected there. Otherwise, if teh op class was declared using
sakura-sandbox, but the operator was instantiated using the GUI,
the standard streams are redirected to the sakura-sandbox terminal.