Rework source column management and related API in operators source
Previously, the "source column" was a vague concept.

For instance, in the source code of an operator, one may have to:
1) filter source0 with a filter on column A => get source1
2) select only column B on source1 => get source2
3) do computations on source2

In this example column A is actually a column of source0 whereas
column B is a column of source1.
However:
- 'column A' also has a meaning in source1
- 'column B' also has a meaning in source0 and source2

Previously, to handle this problem, we were mostly using column
indices. But it was making the operator code less obvious.

In this commit, we make a distinction between a 'Column' and a
'BoundColumn'.
When instanciating a 'Column' object, a uuid is generated, which
allows to identify it in the whole process (e.g. after a filter
operation on the source, the column can still be identified because
its uuid is still there).
The 'BoundColumn' object allows to link a 'Column' to a given source,
and get a stream of values.

This change maintains compatibility with the source code of existing
operators. However, it also provides alternate (more intuitive)
source manipulation methods select() and filter() (see map/operator.py).

Another minimalistic change was done: column selection parameters
now provide a 'column' parameter which is directly bound to the
appropriate 'BoundColumn'. It allows to simplify operator source code
when 'col_index' attribute was used.