Handle approximate chunks and apply to a progressive sort algorithm
NumpyChunk objects have a new 'exactness' attr, having 3 possible values:
- EXACT
- APPROXIMATE
- UNDEFINED

State UNDEFINED should only be a temporary value.
APPROXIMATE chunks allow to yield an approximate preview of the next
EXACT chunk, when processing time since the last emitted chunk is too long.

For instance, let's consider this list of integers:
>>> a = np.arange(10000000, dtype=np.uint32)
>>> a
array([       0,        1,        2, ..., 9999997, 9999998, 9999999],
      dtype=uint32)
>>> random.shuffle(a)
>>> a
array([3554749, 6792485, 6687523, ..., 9582444, 2955532, 5574483],
      dtype=uint32)
>>>

A progressive sort algorithm should be able to yield the following
chunks:
[ 0, 1, 3, 5, 8 ]     APPROXIMATE
[ 0, 1, 2, 3, 5 ]     APPROXIMATE
[ 0, 1, 2, 3, 4 ]     APPROXIMATE (*)
[ 0, 1, 2, 3, 4 ]     EXACT
[ 5, 6, 8, 9, 10 ]    APPROXIMATE
[ 5, 6, 7, 8, 9 ]     EXACT
...
(*) of course, an approximate chunk might actually be exact.

The important thing here is that only an EXACT chunk allows to advance
in the stream of values. Again, an APPROXIMATE chunk is just an
approximation of the next EXACT chunk.

For now we just have a single algorithm that yields approximate
chunks. This is a progressive sort algorithm and it is used when
sorting a computed source.

We preserve compatbility with previous code:
* When the callback of a ChunksComputedSource yields chunks with
  exactness = UNDEFINED, we update it to EXACT. Of course the developer
  may yield chunks with explicit exactness defined.
* the callback of an ItemsComputedSource may yield one of these two
  forms of output:
  yield <row_values>, <exactness>       # new form
  yield <row_values>                    # old form
  with the old form, <row_values> are considered EXACT.
* SourceBase.chunks() prototype was updated to:
  chunks(self, chunk_size = None, allow_approximate = False)
  This calls self.all_chunks() (defined in subclasses) and it discards
  APPROXIMATE chunks if allow_approximate is False. Since this is
  the default value, previous code works as expected.

The mechanism involved in ItemsComputedSource callbacks is similar to
the one involving chunks: a row with flag APPROXIMATE is an
approximation of this row, and only a row marked EXACT allows to advance
to the next one.