Improve technique to quickly get first sql query rows
The technique can only be applied when a sorting is applied.
The fact no sorting is required does not mean a sorted result is invalid.
Thus this commit tries to find a primary key that can be used for the sort,
with a low impact on performance since primary keys always have an index,
and then applies the technique mentioned above.