Operator custom tab: Remove iframe sandbox attribute
With recent chrome browsers, websocket connection apparently does not work
anymore inside an iframe when the sandbox attribute is set.
Since developer grant is needed to develop an operator, only trusted
users can provide js & html code run in such an iframe, so removing the
sandbox attribute should not be a problem.